+++ programs/vrf-program/src/state/random_value.rs
@@ -1,16 +1,12 @@
 use anchor_lang::prelude::*;
 
-/// whitelists are used to control what gems can/can't go into the vault
-/// currently 2 types of vault lists are supported: by mint and by creator
-/// if the whitelist PDA exists, then the mint/creator is considered accepted
-/// if at least 1 whitelist PDA exists total, then all deposit attempts will start getting checked
 #[repr(C)]
 #[account]
 
 pub struct RandomValue {
-    pub min: u64,
-    pub max: u64,
-    pub result: u64,
+    pub min: u32,
+    pub max: u32,
+    pub result: u32,
     pub processed: bool,
 }
